+++ base/boot.jl
@@ -471,6 +471,14 @@ struct UUID
     UUID(value::UInt128) = new(value)
 end
 
+abstract type AbstractLibrary end
+
+struct LibraryID
+    pkg::UUID
+    name::String
+    LibraryID(pkg::UUID, name::String) = new(pkg, name)
+end
+
 abstract type Exception end
 struct ErrorException <: Exception
     msg::AbstractString